refactor: splitting vertexai and aiplatform into separate files
