+++ packages/instrumentation-vertexai/src/types.ts
@@ -7,3 +7,11 @@ export interface VertexAIInstrumentationConfig extends InstrumentationConfig {
    */
   traceContent?: boolean;
 }
+
+export interface AIPlatformInstrumentationConfig extends InstrumentationConfig {
+  /**
+   * Whether to log prompts, completions and embeddings on traces.
+   * @default true
+   */
+  traceContent?: boolean;
+}
